requirements.txt
google-analytics-data
klaviyo-api
openai
pandas
ShopifyAPI
streamlit